Audit fixes (#81)
* Add controller for minting and creating accounts atomically

* Generalize controller

* Fix requirePermits logic

* Remove erc6551 dependency

* Rename and reorganize

* :art: Add account return parameter and MintParams struct

* :white_check_mark: Add test file for new MintCreateInitializeController

* :sparkles: Use new mintStartTokenId return param from ERC721Rails::mintTo() for mintAndCreateAccount()

* :white_check_mark: Bring changes through test files, point at correct 0xrails commit

* :lock: Say-GR-07 & 08: misleading comments, gas optimizations

* :lock: Say-GR_06: payable removal

* :lock: Comment/remove unused variable names

* :lock: Say-GR-04: Restrict StablecoinPurchaserController mint functions payability, update names

* :lock: Say-GR-03 Prevent empty transfers to payoutAddress/address(0) by _collectFeeAndForwardCollectionRevenue when called in FreeMintController context

* :lock: Say-Gr-02: Use 0xrails::Ownable for FeeController, update tests

* :lock: Say-Gr-05: Add whitelist of valid token implementations to storage to prevent token spoofing and reputation fallout

* :lock: Say-Sa-01: Precompute storage slot for gas optimization and update to newest 7201 spec

* :lock: Say-Gr-01: Mitigate cases of stablecoin depeg by introducing pausable functionality

* :sparkles: Maximize backwards-compatibility and cross-chain convenience of TokenFactory trusted token implementations using storage arrays

---------

+++ test/membership/guards/OnePerAddress.t.sol
@@ -38,7 +38,7 @@ contract OnePerAddressGuardTest is Test, SetUpMembership {
         assertEq(proxy.totalSupply(), 1);
     }
 
-    function test_mintRevertOnePerAddress(uint16 quantity) public {
+    function test_mintRevertOnePerAddress(uint8 quantity) public {
         address to = owner;
         vm.assume(to != address(0));
         vm.assume(quantity > 1);
